Kubernetes controller namespace is Medulla.Kubernetes.Operator controllers should be in the Medulla.Kubernetes.Operator.Controllers namespace.
