--- src/Controller/.gitignore
@@ -1,1 +1,0 @@
-﻿
