--- src/Controller/build.ps1
@@ -1,1 +1,0 @@
-﻿sudo dotnet run -- generator docker --solution-dir .. --project-path DatabaseControllerKubeOps.csproj -o ./Dockerfile
