--- src/Controller/config/crds/data_medulla_recro_com.yaml
@@ -1,153 +1,0 @@
-apiVersion: apiextensions.k8s.io/v1
-kind: CustomResourceDefinition
-metadata:
-  name: data.medulla.recro.com
-spec:
-  group: medulla.recro.com
-  names:
-    kind: Data
-    listKind: DataList
-    plural: data
-    singular: data
-  scope: Namespaced
-  versions:
-  - name: v1alpha1
-    schema:
-      openAPIV3Schema:
-        description: A CustomResourceDefinition which allows building with data in Medulla.
-        properties:
-          spec:
-            items:
-              properties:
-                name:
-                  nullable: true
-                  type: string
-                host:
-                  nullable: true
-                  type: string
-                dialect:
-                  nullable: true
-                  type: string
-                usernameSecretName:
-                  nullable: true
-                  type: string
-                passwordSecretName:
-                  nullable: true
-                  type: string
-                models:
-                  items:
-                    properties:
-                      name:
-                        nullable: true
-                        type: string
-                      columns:
-                        items:
-                          properties:
-                            columnName:
-                              nullable: true
-                              type: string
-                            type:
-                              nullable: true
-                              type: string
-                            allowNull:
-                              type: boolean
-                            defaultValue:
-                              nullable: true
-                              type: string
-                            autoIncrement:
-                              type: boolean
-                            primaryKey:
-                              type: boolean
-                            field:
-                              nullable: true
-                              type: string
-                            unique:
-                              nullable: true
-                              type: string
-                            comment:
-                              nullable: true
-                              type: string
-                            validate:
-                              nullable: true
-                              properties:
-                                hasValue:
-                                  type: boolean
-                                value:
-                                  properties:
-                                    is:
-                                      nullable: true
-                                      type: string
-                                    not:
-                                      nullable: true
-                                      type: string
-                                    isEmail:
-                                      nullable: true
-                                      type: boolean
-                                    isUrl:
-                                      nullable: true
-                                      type: boolean
-                                    isIp:
-                                      nullable: true
-                                      type: boolean
-                                    isIpV4:
-                                      nullable: true
-                                      type: boolean
-                                    isIpV6:
-                                      nullable: true
-                                      type: boolean
-                                    isAlpha:
-                                      nullable: true
-                                      type: boolean
-                                    isAlphaNumeric:
-                                      nullable: true
-                                      type: boolean
-                                    isNumeric:
-                                      nullable: true
-                                      type: boolean
-                                    isInt:
-                                      nullable: true
-                                      type: boolean
-                                    isFloat:
-                                      nullable: true
-                                      type: boolean
-                                    isDecimal:
-                                      nullable: true
-                                      type: boolean
-                                    isLowercase:
-                                      nullable: true
-                                      type: boolean
-                                    isUppercase:
-                                      nullable: true
-                                      type: boolean
-                                    isNull:
-                                      nullable: true
-                                      type: boolean
-                                    notNull:
-                                      nullable: true
-                                      type: boolean
-                                    notEmpty:
-                                      nullable: true
-                                      type: boolean
-                                  type: object
-                              type: object
-                          type: object
-                        nullable: true
-                        type: array
-                    required:
-                    - name
-                    - columns
-                    type: object
-                  type: array
-              required:
-              - name
-              - host
-              - dialect
-              - usernameSecretName
-              - passwordSecretName
-              - models
-              type: object
-            nullable: true
-            type: array
-        type: object
-    served: true
-    storage: true
