--- src/Controller/config/crds/kustomization.yaml
@@ -1,6 +1,0 @@
-commonLabels:
-  operator-element: crd
-resources:
-- data_medulla_recro_com.yaml
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
