--- src/Controller/config/install/kustomization.yaml
@@ -1,15 +1,0 @@
-namespace: databasecontrollerkubeops-system
-namePrefix: databasecontrollerkubeops-
-commonLabels:
-  operator: databasecontrollerkubeops
-resources:
-- ./namespace.yaml
-- ../crds
-- ../rbac
-- ../operator
-images:
-- name: operator
-  newName: keithlogan94/controller
-  newTag: latest
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
