--- src/Controller/config/operator/deployment.yaml
@@ -1,51 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  labels:
-    operator-deployment: databasecontrollerkubeops
-  name: operator
-spec:
-  replicas: 1
-  revisionHistoryLimit: 0
-  selector:
-    matchLabels:
-      operator: databasecontrollerkubeops
-  template:
-    metadata:
-      labels:
-        operator: databasecontrollerkubeops
-    spec:
-      containers:
-      - env:
-        - name: POD_NAMESPACE
-          valueFrom:
-            fieldRef:
-              fieldPath: metadata.namespace
-        envFrom:
-        - configMapRef:
-            name: webhook-config
-        image: operator
-        livenessProbe:
-          httpGet:
-            path: /health
-            port: http
-          initialDelaySeconds: 30
-          timeoutSeconds: 1
-        name: operator
-        ports:
-        - containerPort: 5000
-          name: http
-        readinessProbe:
-          httpGet:
-            path: /ready
-            port: http
-          initialDelaySeconds: 15
-          timeoutSeconds: 1
-        resources:
-          limits:
-            cpu: 100m
-            memory: 128Mi
-          requests:
-            cpu: 100m
-            memory: 64Mi
-      terminationGracePeriodSeconds: 10
