--- src/Controller/config/operator/kustomization.yaml
@@ -1,10 +1,0 @@
-commonLabels:
-  operator-element: operator-instance
-resources:
-- deployment.yaml
-configMapGenerator:
-- name: webhook-config
-  literals:
-  - KESTREL__ENDPOINTS__HTTP__URL=http://0.0.0.0:5000
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
