--- src/Controller/config/rbac/kustomization.yaml
@@ -1,7 +1,0 @@
-commonLabels:
-  operator-element: rbac
-resources:
-- operator-role.yaml
-- operator-role-binding.yaml
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
