--- src/Controller/config/rbac/operator-role-binding.yaml
@@ -1,12 +1,0 @@
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: operator-role-binding
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: operator-role
-subjects:
-- kind: ServiceAccount
-  name: default
-  namespace: system
