--- src/Controller/config/rbac/operator-role.yaml
@@ -1,37 +1,0 @@
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  name: operator-role
-rules:
-- apiGroups:
-  - medulla.recro.com
-  - coordination.k8s.io
-  resources:
-  - data
-  - leases
-  verbs:
-  - '*'
-- apiGroups:
-  - ''
-  resources:
-  - events
-  verbs:
-  - create
-  - get
-  - list
-  - update
-- apiGroups:
-  - apps
-  resources:
-  - deployments
-  verbs:
-  - get
-  - list
-- apiGroups:
-  - apps
-  resources:
-  - deployments/status
-  verbs:
-  - get
-  - patch
-  - update
