--- src/Controller/test/medulla-data.yaml
@@ -1,116 +1,0 @@
-﻿# Licensed to the Medulla Contributors under one or more agreements.
-# The Medulla Contributors licenses this file to you under the Apache 2.0 license.
-# See the LICENSE file in the project root for more information.
-
-apiVersion: medulla.recro.com/v1alpha1
-kind: Data
-metadata:
-  name: test-1
-databases:
-  - name: medulla
-    host: mysql-service
-    dialect: mysql
-    usernameSecretName: user
-    passwordSecretName: test
-    models:
-      - name: User
-        columns:
-          - columnName: email
-            type: VARCHAR(250)
-            allowNull: true
-            defaultValue: "test@test.com"
-            primaryKey: true
-            field: "email"
-            unique: "email"
-            comment: "This is the email field"
-      - name: Account
-        columns:
-          - columnName: phoneNumber
-            type: VARCHAR(50)
-            allowNull: true
-            defaultValue: ""
-            primaryKey: false
-            field: "phone_number"
-            unique: "uniqueField"
-            comment: "Stores the phone number for the user"
-
-      - name: Products
-        columns:
-          - columnName: productId
-            type: INT(11)
-            allowNull: false
-            defaultValue: "1"
-            primaryKey: true
-            field: "product_id"
-            unique: "uniqueField"
-            comment: "Product id for Product"
-
-      - name: Files
-        columns:
-          - columnName: fileName
-            type: VARCHAR(100)
-            allowNull: false
-            defaultValue: ""
-            primaryKey: true
-            field: "file_name"
-            unique: "uniqueField"
-            comment: "File"
-        
-
-      
-      - name: TestTable
-        columns:
-          - columnName: testColumn
-            type: VARCHAR(100)
-            allowNull: false
-            defaultValue: ""
-            primaryKey: true
-            field: "test_column"
-            unique: "test"
-            comment: "This is a test"
-          
-          - columnName: phoneNumber
-            type: VARCHAR(100)
-            allowNull: true
-            defaultValue: "555-555-5555"
-            primaryKey: false
-            field: "phone_number"
-            unique: "phoneNumber"
-            comment: "This is a phone number field"
-
-          - columnName: phoneNumber0
-            type: VARCHAR(100)
-            allowNull: true
-            defaultValue: "555-555-5555"
-            primaryKey: false
-            field: "phone_number0"
-            unique: "phoneNumber"
-            comment: "This is a phone number field"
-
-          - columnName: phoneNumber1
-            type: VARCHAR(100)
-            allowNull: true
-            defaultValue: "555-555-5555"
-            primaryKey: false
-            field: "phone_number1"
-            unique: "phoneNumber"
-            comment: "This is a phone number field"
-
-          - columnName: phoneNumber2
-            type: VARCHAR(100)
-            allowNull: true
-            defaultValue: "555-555-5555"
-            primaryKey: false
-            field: "phone_number2"
-            unique: "phoneNumber"
-            comment: "This is a phone number field"
-
-          - columnName: phoneNumber3
-            type: VARCHAR(100)
-            allowNull: true
-            defaultValue: "555-555-5555"
-            primaryKey: false
-            field: "phone_number3"
-            unique: "phoneNumber"
-            comment: "This is a phone number field"
-
