+++ src/Medulla/Kubernetes/Operator/appsettings.Development.json
