+++ src/Medulla/Kubernetes/Operator/appsettings.json
